Added assignee search methods in addition to grant search methods Program gets hung up after several minutes

+++ manual_search.py
@@ -11,17 +11,107 @@ import os
 import pprint
 
 
-def govt_search(query):
-    directory = "grant search results html"
+def assignee_search(query):
+    directory = "assignee search results"
     if not os.path.exists(directory):
         os.makedirs(directory)
 
-    filename = "grant search results html/" + query.replace("/", "") + ".html"
+    filename = "assignee search results/" + query.replace(" ", "_").replace("/", "_") + ".html"
+    print query
+
+    if not os.path.exists(filename):
+        data = {"Sect1": "PTO2",
+                "Sect2": "HITOFF",
+                "u": "/netahtml/PTO/search-adv.htm",
+                "r": "0",
+                "p": "1",
+                "f": "S",
+                "l": "50",
+                "Query": 'AN/"' + query + '"',
+                "d": "PTXT"}
+        r = requests.get("http://patft.uspto.gov/netacgi/nph-Parser", data)
+        with open(filename, 'w') as f:
+            f.write(r.text)
+        print "Saved file: " + filename
+
+        time.sleep(0.5)
+    return filename
+
+
+# Used when there are multiple pages of results to be named _2, _3, etc.
+def assignee_search_next(query, page):
+    directory = "assignee search results"
+    if not os.path.exists(directory):
+        os.makedirs(directory)
+
+    filename = "assignee search results/" + query.replace(" ", "_").replace("/", "_") + "_" + str(page) + ".html"
+    print query
+
+    if not os.path.exists(filename):
+        data = {"Sect1": "PTO2",
+                "Sect2": "HITOFF",
+                "u": "/netahtml/PTO/search-adv.htm",
+                "r": "0",
+                "p": page,
+                "f": "S",
+                "l": "50",
+                "OS": 'AN/"' + query + '"',
+                "RS": 'AN/"' + query + '"',
+                "Page": "Next",
+                "S1": query + ".ASNM.",
+                "d": "PTXT"}
+        r = requests.get("http://patft.uspto.gov/netacgi/nph-Parser", data)
+        with open(filename, 'w') as f:
+            f.write(r.text)
+        print "Saved file: " + filename
+
+        time.sleep(0.5)
+    return filename
+
+
+# Used when there is only one patent and the search page redirects
+def assignee_search_1(query):
+    directory = "assignee search results"
+    if not os.path.exists(directory):
+        os.makedirs(directory)
+
+    filename = "assignee search results/" + query.replace(" ", "_").replace("/", "_") + "_1.html"
+    query = '("' + query + '".ASNM.)'
+    print query
+
+    if not os.path.exists(filename):
+        data = {"Sect1": "PTO2",
+                "Sect2": "HITOFF",
+                "u": "/netahtml/PTO/search-adv.htm",
+                "r": "1",
+                "p": "1",
+                "f": "G",
+                "l": "50",
+                "S1": query,
+                "d": "PTXT"}
+        r = requests.get("http://patft.uspto.gov/netacgi/nph-Parser", data)
+        print r.url
+        with open(filename, 'w') as f:
+            f.write(r.text)
+        print "Saved file: " + filename
+
+        time.sleep(0.5)
+    return filename
+
+
+def govt_search(query):
+    directory = "grant search results"
+    if not os.path.exists(directory):
+        os.makedirs(directory)
+
+    filename = "grant search results/" + query.replace("/", "") + ".html"
     if query[0:2] == "DE":
-        number = query[5:] # e.g. 000032
-        office_code = query[3:5] # e.g. AR
-        query = "GOVT/" + office_code + number + ' OR GOVT/"' + \
-                office_code + " " + number + '"'
+        number = query[5:]  # e.g. 000032
+        office_code = query[3:5]  # e.g. AR
+        query = "GOVT/" + office_code + number + \
+                '$ OR GOVT/"' + office_code + " " + number + '"$' + \
+                " OR GOVT/" + office_code + number + \
+                ' OR GOVT/"' + office_code + " " + number + '"'
         print query
     else:
         query = "GOVT/" + query
@@ -45,18 +135,62 @@ def govt_search(query):
     return filename
 
 
-# Used when there is only one citing patent and the search page redirects
-def govt_search_1(query):
-    directory = "grant search results html"
+# Used when there are multiple pages of results to be named _2, _3, etc.
+# THIS NEEDS TO BE CHECKED USING AN ACTUAL EXAMPLE OF A MULTIPAGE GRANT SEARCH RESULT
+def govt_search_next(query, page):
+    directory = "grant search results"
     if not os.path.exists(directory):
         os.makedirs(directory)
 
-    filename = "grant search results html/" + query.replace("/", "") + "_1.html"
+    filename = "grant search results/" + query.replace("/", "") + "_" + str(page) + ".html"
+    if query[0:2] == "DE":
+        number = query[5:]  # e.g. 000032
+        office_code = query[3:5]  # e.g. AR
+        query = "GOVT/" + office_code + number + \
+                '$ OR GOVT/"' + office_code + " " + number + '"$' + \
+                " OR GOVT/" + office_code + number + \
+                ' OR GOVT/"' + office_code + " " + number + '"'
+        print query
+    else:
+        query = "GOVT/" + query
+
+    if not os.path.exists(filename):
+        data = {"Sect1": "PTO2",
+                "Sect2": "HITOFF",
+                "u": "/netahtml/PTO/search-adv.htm",
+                "r": "0",
+                "p": page,
+                "f": "S",
+                "l": "50",
+                "OS": query,
+                "RS": query,
+                "Query": query,
+                "Page": "Next",
+                "S1": query + ".GOVT.",
+                "d": "PTXT"}
+        r = requests.get("http://patft.uspto.gov/netacgi/nph-Parser", data)
+        with open(filename, 'w') as f:
+            f.write(r.text)
+        print "Saved file: " + filename
+
+        time.sleep(0.5)
+    return filename
+
+
+# Used when there is only one patent and the search page redirects
+def govt_search_1(query):
+    directory = "grant search results"
+    if not os.path.exists(directory):
+        os.makedirs(directory)
+
+    filename = "grant search results/" + query.replace("/", "") + "_1.html"
     if query[0:2] == "DE":
         number = query[5:] # e.g. 000032
         office_code = query[3:5] # e.g. AR
-        query = "(" + office_code + number + '.GOTX. OR ("' + \
-                office_code + " " + number + '".GOTX.))'
+        query = "(" + office_code + number + '$.GOTX. OR "' + \
+                office_code + " " + number + '"$.GOTX. OR ' + \
+                office_code + number + '.GOTX. OR "' + \
+                office_code + " " + number + '".GOTX.)'
         print query
     else:
         query = "GOVT/" + query
@@ -71,38 +205,6 @@ def govt_search_1(query):
                 "l": "50",
                 "S1": query,
                 "d": "PTXT"}
-        r = requests.get("http://patft.uspto.gov/netacgi/nph-Parser", data)
-        with open(filename, 'w') as f:
-            f.write(r.text)
-        print "Saved file: " + filename
-
-        time.sleep(0.5)
-    return filename
-
-def pn_search(query, type):
-    # Type is "citing" or "DOE"-- determines which folder the patent is stored in
-    if type == "DOE":
-        filename = "patent pages html/" + query + ".html"
-        directory = "patent pages html"
-    else:
-        filename = "citing patents html/" + query + ".html"
-        directory = "citing patents html"
-
-    if not os.path.exists(directory):
-        os.makedirs(directory)
-
-    if not os.path.exists(filename):
-        data = {"Sect1": "PTO2",
-                "Sect2": "HITOFF",
-                "u": "/netahtml/PTO/search-adv.htm",
-                "r": "1",
-                "p": "1",
-                "f": "G",
-                "l": "50",
-                "d": "PTXT",
-                "S1": str(query) + ".PN.",
-                "OS": "PN/" + str(query),
-                "RS": "PN/" + str(query)}
         r = requests.get("http://patft.uspto.gov/netacgi/nph-Parser", data)
         with open(filename, 'w') as f:
             f.write(r.text)
@@ -127,6 +229,35 @@ def ref_search(query):
                 "f": "S",
                 "l": "50",
                 "Query": "REF/" + str(query),
+                "d": "PTXT"}
+        r = requests.get("http://patft.uspto.gov/netacgi/nph-Parser", data)
+        with open(filename, 'w') as f:
+            f.write(r.text)
+        print "Saved file: " + filename
+
+        time.sleep(0.5)
+    return filename
+
+
+# Used when there are multiple pages of results to be named _2, _3, etc.
+def ref_search_next(query, page):
+    directory = "reference search results html"
+    if not os.path.exists(directory):
+        os.makedirs(directory)
+
+    filename = "reference search results html/" + query + "_" + str(page) + ".html"
+    if not os.path.exists(filename):
+        data = {"Sect1": "PTO2",
+                "Sect2": "HITOFF",
+                "u": "/netahtml/PTO/search-adv.htm",
+                "r": "0",
+                "p": page,
+                "f": "S",
+                "l": "50",
+                "OS": "REF/" + str(query),
+                "RS": "REF/" + str(query),
+                "Page": "Next",
+                "S1": str(query) + ".UREF.",
                 "d": "PTXT"}
         r = requests.get("http://patft.uspto.gov/netacgi/nph-Parser", data)
         with open(filename, 'w') as f:
@@ -165,27 +296,44 @@ def ref_search_1(query):
     return filename
 
 
-# Used when there is more than one page of citing patents
-def ref_search_2(query):
-    directory = "reference search results html"
+def run_search(csv_file, search_unit):
+    with open(csv_file, "rb") as f:
+        text = csv.reader(f)
+        queries = [row[0] for row in text]
+        files = []
+        for x in queries:
+            if search_unit == "grant":
+                filename = govt_search(x)
+            elif search_unit == "assignee":
+                filename = assignee_search(x)
+            files.append(filename)
+    return [queries, files]
+
+
+def pn_search(query, type):
+    # Type is "citing" or "primary"-- determines which folder the patent is stored in
+    if type == "primary":
+        filename = "patent pages html/" + query + ".html"
+        directory = "patent pages html"
+    else:
+        filename = "citing patents html/" + query + ".html"
+        directory = "citing patents html"
+
     if not os.path.exists(directory):
         os.makedirs(directory)
 
-    filename = "reference search results html/" + query + "_2.html"
     if not os.path.exists(filename):
         data = {"Sect1": "PTO2",
                 "Sect2": "HITOFF",
                 "u": "/netahtml/PTO/search-adv.htm",
-                "r": "0",
-                "f": "S",
+                "r": "1",
+                "p": "1",
+                "f": "G",
                 "l": "50",
                 "d": "PTXT",
-                "OS": "REF/" + str(query),
-                "RS": "REF/" + str(query),
-                "Query": "REF/" + str(query),
-                "TD": "1",
-                "Srch1": str(query) + ".UREF.",
-                "NextList2": "Final"}
+                "S1": str(query) + ".PN.",
+                "OS": "PN/" + str(query),
+                "RS": "PN/" + str(query)}
         r = requests.get("http://patft.uspto.gov/netacgi/nph-Parser", data)
         with open(filename, 'w') as f:
             f.write(r.text)
@@ -195,18 +343,25 @@ def ref_search_2(query):
     return filename
 
 
-def run_search(csv_file):
-    with open(csv_file, "rb") as f:
-        text = csv.reader(f)
-        grants = [row[0] for row in text]
-        files = []
-        for x in grants:
-            filename = govt_search(x)
-            files.append(filename)
-    return [grants, files]
+def single_result_search(query, search_unit):
+    if search_unit == "grant":
+        filename = govt_search_1(query)
+    elif search_unit == "assignee":
+        filename = assignee_search_1(query)
+
+    return filename
 
 
-csv_file = "DOE grant short list.csv"
+def multiple_page_search(query, search_unit, page):
+    if search_unit == "grant":
+        filename = govt_search_next(query, page)
+    elif search_unit == "assignee":
+        filename = assignee_search_next(query, page)
+
+    return filename
+
+
+# csv_file = "DOE grant short list.csv"
 
 if __name__ == '__main__':
     # pprint.pprint(run_search(csv_file))
@@ -216,7 +371,9 @@ if __name__ == '__main__':
     # ref_search("8047593")
     # patent cited 31 times
     # ref_search("8307899")
-    govt_search_1("DE-SC0003254")
+    # assignee_search("1366 TECHNOLOGIES")
+    # run_search("test assignees.csv", "assignee")
+    single_result_search("ABEAM TECHNOLOGIES", "assignee")
